+++ R/predict_functions.R
@@ -49,8 +49,7 @@ predict_functions <- function(name) {
     "glmer_binomial",
     "randomForest_gaussian"
   )) {
-    predict_fn <- function(test_data, model, formula,
-                           hyperparameters, train_data) {
+    predict_fn <- function(test_data, model, formula, hyperparameters, train_data) {
       stats::predict(
         object = model,
         newdata = test_data,
@@ -59,8 +58,7 @@ predict_functions <- function(name) {
       )
     }
   } else if (name %in% c("svm_gaussian")) {
-    predict_fn <- function(test_data, model, formula,
-                           hyperparameters, train_data) {
+    predict_fn <- function(test_data, model, formula, hyperparameters, train_data) {
       stats::predict(
         object = model,
         newdata = test_data,
@@ -68,8 +66,7 @@ predict_functions <- function(name) {
       )
     }
   } else if (name == "svm_binomial") {
-    predict_fn <- function(test_data, model, formula,
-                           hyperparameters, train_data) {
+    predict_fn <- function(test_data, model, formula, hyperparameters, train_data) {
       predictions <- stats::predict(
         object = model,
         newdata = test_data,
@@ -86,8 +83,7 @@ predict_functions <- function(name) {
       probabilities[[2]]
     }
   } else if (name == "svm_multinomial") {
-    predict_fn <- function(test_data, model, formula,
-                           hyperparameters, train_data) {
+    predict_fn <- function(test_data, model, formula, hyperparameters, train_data) {
       predictions <- stats::predict(
         object = model,
         newdata = test_data,
@@ -104,8 +100,7 @@ predict_functions <- function(name) {
       probabilities
     }
   } else if (name == "naive_bayes") {
-    predict_fn <- function(test_data, model, formula,
-                           hyperparameters, train_data) {
+    predict_fn <- function(test_data, model, formula, hyperparameters, train_data) {
       stats::predict(
         object = model,
         newdata = test_data,
@@ -114,8 +109,7 @@ predict_functions <- function(name) {
       )[, 2]
     }
   } else if (name == "nnet_multinom") {
-    predict_fn <- function(test_data, model, formula,
-                           hyperparameters, train_data) {
+    predict_fn <- function(test_data, model, formula, hyperparameters, train_data) {
       stats::predict(
         object = model,
         newdata = test_data,
@@ -124,8 +118,7 @@ predict_functions <- function(name) {
       )
     }
   } else if (name %in% c("nnet_gaussian", "nnet_binomial")) {
-    predict_fn <- function(test_data, model, formula,
-                           hyperparameters, train_data) {
+    predict_fn <- function(test_data, model, formula, hyperparameters, train_data) {
       stats::predict(
         object = model,
         newdata = test_data,
@@ -134,8 +127,7 @@ predict_functions <- function(name) {
       )
     }
   } else if (name %in% c("randomForest_binomial")) {
-    predict_fn <- function(test_data, model, formula,
-                           hyperparameters, train_data) {
+    predict_fn <- function(test_data, model, formula, hyperparameters, train_data) {
       stats::predict(
         object = model,
         newdata = test_data,
@@ -143,8 +135,7 @@ predict_functions <- function(name) {
       )[, 2]
     }
   } else if (name %in% c("randomForest_multinomial")) {
-    predict_fn <- function(test_data, model, formula,
-                           hyperparameters, train_data) {
+    predict_fn <- function(test_data, model, formula, hyperparameters, train_data) {
       stats::predict(
         object = model,
         newdata = test_data,